+++ tybotRoute/test/single_test.sh
@@ -2,4 +2,4 @@
 #npx mocha ./test/mock_query_test.js
 #npx mocha ./test/disable_input_text_directive_test.js
 #npx mocha ./test/close_directive_test.js
-npx mocha ./test/conversation1-test.js -r dotenv/config
+npx mocha ./test/conversation1-test.js --exit